services/orders/requirements.txt
fastapi
uvicorn
python-dotenv
aiokafka
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi